requirements.txt
Flask
google-generativeai==0.7.1
Flask-Cors